# use thomaswue trick, use parallelism, slightly faster