@gunnarmorling 2024-01-23
00:02.305 now, nice!